Refactor Subprocesso services into unified SubprocessoService
