--- backend/src/main/java/sgc/subprocesso/service/SubprocessoMapaWorkflowService.java
@@ -1,425 +1,0 @@
-package sgc.subprocesso.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.jspecify.annotations.Nullable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
-import sgc.comum.erros.ErroValidacao;
-import sgc.mapa.MapaFacade;
-import sgc.mapa.dto.SalvarMapaRequest;
-import sgc.mapa.model.Atividade;
-import sgc.mapa.model.Mapa;
-import sgc.mapa.service.MapaManutencaoService;
-import sgc.organizacao.OrganizacaoFacade;
-import sgc.organizacao.model.Unidade;
-import sgc.organizacao.model.Usuario;
-import sgc.processo.model.TipoProcesso;
-import sgc.subprocesso.AnaliseFacade;
-import sgc.subprocesso.dto.*;
-import sgc.subprocesso.model.*;
-
-import java.time.LocalDateTime;
-import java.util.EnumMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import static sgc.subprocesso.model.SituacaoSubprocesso.*;
-
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class SubprocessoMapaWorkflowService {
-
-    private static final String SIGLA_ADMIN = "ADMIN";
-
-    private static final Map<TipoProcesso, SituacaoSubprocesso> SITUACAO_MAPA_DISPONIBILIZADO = new EnumMap<>(Map.of(
-            TipoProcesso.MAPEAMENTO, MAPEAMENTO_MAPA_DISPONIBILIZADO,
-            TipoProcesso.REVISAO, REVISAO_MAPA_DISPONIBILIZADO));
-
-    private static final Map<TipoProcesso, SituacaoSubprocesso> SITUACAO_MAPA_COM_SUGESTOES = new EnumMap<>(Map.of(
-            TipoProcesso.MAPEAMENTO, MAPEAMENTO_MAPA_COM_SUGESTOES,
-            TipoProcesso.REVISAO, REVISAO_MAPA_COM_SUGESTOES));
-
-    private static final Map<TipoProcesso, SituacaoSubprocesso> SITUACAO_MAPA_VALIDADO = new EnumMap<>(Map.of(
-            TipoProcesso.MAPEAMENTO, MAPEAMENTO_MAPA_VALIDADO,
-            TipoProcesso.REVISAO, REVISAO_MAPA_VALIDADO));
-
-    private static final Map<TipoProcesso, SituacaoSubprocesso> SITUACAO_MAPA_HOMOLOGADO = new EnumMap<>(Map.of(
-            TipoProcesso.MAPEAMENTO, MAPEAMENTO_MAPA_HOMOLOGADO,
-            TipoProcesso.REVISAO, REVISAO_MAPA_HOMOLOGADO));
-
-    private final SubprocessoRepo subprocessoRepo;
-    private final MovimentacaoRepo movimentacaoRepo;
-    private final SubprocessoWorkflowService workflowService;
-    private final MapaManutencaoService mapaManutencaoService;
-    private final MapaFacade mapaFacade;
-    private final AnaliseFacade analiseFacade;
-    private final SubprocessoTransicaoService transicaoService;
-    private final OrganizacaoFacade organizacaoFacade;
-
-    public Mapa salvarMapaSubprocesso(Long codSubprocesso, SalvarMapaRequest request) {
-        Subprocesso subprocesso = getSubprocessoParaEdicao(codSubprocesso);
-        Long codMapa = subprocesso.getMapa().getCodigo();
-        boolean eraVazio = mapaManutencaoService.buscarCompetenciasPorCodMapa(codMapa).isEmpty();
-        boolean temNovasCompetencias = !request.competencias().isEmpty();
-
-        Mapa mapa = mapaFacade.salvarMapaCompleto(codMapa, request);
-        if (eraVazio && temNovasCompetencias) {
-            if (subprocesso.getSituacao() == MAPEAMENTO_CADASTRO_HOMOLOGADO) {
-                subprocesso.setSituacao(MAPEAMENTO_MAPA_CRIADO);
-                subprocessoRepo.save(subprocesso);
-            } else if (subprocesso.getSituacao() == REVISAO_CADASTRO_HOMOLOGADA) {
-                subprocesso.setSituacao(REVISAO_MAPA_AJUSTADO);
-                subprocessoRepo.save(subprocesso);
-            }
-        }
-
-        return mapa;
-    }
-
-    public Mapa adicionarCompetencia(Long codSubprocesso, CompetenciaRequest request) {
-        Subprocesso subprocesso = getSubprocessoParaEdicao(codSubprocesso);
-        Mapa mapa = subprocesso.getMapa();
-
-        Long codMapa = mapa.getCodigo();
-        boolean eraVazio = mapaManutencaoService.buscarCompetenciasPorCodMapa(codMapa).isEmpty();
-
-        mapaManutencaoService.criarCompetenciaComAtividades(
-                mapa, request.descricao(), request.atividadesIds());
-
-        // Alterar situação para MAPA_CRIADO/AJUSTADO se era vazio e passou a ter competências
-        if (eraVazio) {
-            if (subprocesso.getSituacao() == MAPEAMENTO_CADASTRO_HOMOLOGADO) {
-                subprocesso.setSituacao(MAPEAMENTO_MAPA_CRIADO);
-                subprocessoRepo.save(subprocesso);
-            } else if (subprocesso.getSituacao() == REVISAO_CADASTRO_HOMOLOGADA) {
-                subprocesso.setSituacao(REVISAO_MAPA_AJUSTADO);
-                subprocessoRepo.save(subprocesso);
-            }
-        }
-
-        return mapaFacade.mapaPorCodigo(mapa.getCodigo());
-    }
-
-    public Mapa atualizarCompetencia(
-            Long codSubprocesso,
-            Long codCompetencia,
-            CompetenciaRequest request) {
-
-        Subprocesso subprocesso = getSubprocessoParaEdicao(codSubprocesso);
-        mapaManutencaoService.atualizarCompetencia(codCompetencia, request.descricao(), request.atividadesIds());
-
-        Mapa mapa = subprocesso.getMapa();
-        return mapaFacade.mapaPorCodigo(mapa.getCodigo());
-    }
-
-    public Mapa removerCompetencia(Long codSubprocesso, Long codCompetencia) {
-        Subprocesso subprocesso = getSubprocessoParaEdicao(codSubprocesso);
-
-        Long codMapa = subprocesso.getMapa().getCodigo();
-        mapaManutencaoService.removerCompetencia(codCompetencia);
-
-        // Se o mapa ficou vazio, voltar para situação anterior
-        boolean ficouVazio = mapaManutencaoService.buscarCompetenciasPorCodMapa(codMapa).isEmpty();
-        if (ficouVazio) {
-            if (subprocesso.getSituacao() == MAPEAMENTO_MAPA_CRIADO) {
-                subprocesso.setSituacao(MAPEAMENTO_CADASTRO_HOMOLOGADO);
-                subprocessoRepo.save(subprocesso);
-                log.info("Situação do subprocesso {} alterada para CADASTRO_HOMOLOGADO (mapa ficou vazio)", codSubprocesso);
-            } else if (subprocesso.getSituacao() == REVISAO_MAPA_AJUSTADO) {
-                subprocesso.setSituacao(REVISAO_CADASTRO_HOMOLOGADA);
-                subprocessoRepo.save(subprocesso);
-                log.info("Situação do subprocesso {} alterada para REVISAO_CADASTRO_HOMOLOGADA (mapa ficou vazio)", codSubprocesso);
-            }
-        }
-
-        return mapaFacade.mapaPorCodigo(subprocesso.getMapa().getCodigo());
-    }
-
-    private Subprocesso getSubprocessoParaEdicao(Long codSubprocesso) {
-        Subprocesso subprocesso = workflowService.buscarSubprocesso(codSubprocesso);
-
-        workflowService.validarSituacaoPermitida(subprocesso,
-            "Mapa só pode ser editado com cadastro homologado ou mapa criado. Situação atual: %s".formatted(subprocesso.getSituacao()),
-            MAPEAMENTO_CADASTRO_HOMOLOGADO,
-            MAPEAMENTO_MAPA_CRIADO,
-            REVISAO_CADASTRO_HOMOLOGADA,
-            REVISAO_MAPA_AJUSTADO);
-
-        return subprocesso;
-    }
-
-    @Transactional
-    public void disponibilizarMapa(Long codSubprocesso, DisponibilizarMapaRequest request, Usuario usuario) {
-        executarDisponibilizacaoMapa(codSubprocesso, request, usuario);
-    }
-
-    private void executarDisponibilizacaoMapa(Long codSubprocesso, DisponibilizarMapaRequest request, Usuario usuario) {
-        Subprocesso sp = getSubprocessoParaEdicao(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp,
-                MAPEAMENTO_CADASTRO_HOMOLOGADO, MAPEAMENTO_MAPA_CRIADO, MAPEAMENTO_MAPA_COM_SUGESTOES,
-                REVISAO_CADASTRO_HOMOLOGADA, REVISAO_MAPA_AJUSTADO, REVISAO_MAPA_COM_SUGESTOES);
-
-        validarMapaParaDisponibilizacao(sp);
-        workflowService.validarAssociacoesMapa(sp.getMapa().getCodigo());
-
-        sp.getMapa().setSugestoes(null);
-        analiseFacade.removerPorSubprocesso(codSubprocesso);
-
-        if (StringUtils.hasText(request.observacoes())) {
-            sp.getMapa().setSugestoes(request.observacoes());
-        }
-
-        sp.setSituacao(SITUACAO_MAPA_DISPONIBILIZADO.get(sp.getProcesso().getTipo()));
-
-        sp.setDataLimiteEtapa2(request.dataLimite().atStartOfDay());
-        sp.setDataFimEtapa1(LocalDateTime.now());
-        subprocessoRepo.save(sp);
-
-        Unidade admin = organizacaoFacade.buscarEntidadePorSigla(SIGLA_ADMIN);
-
-        transicaoService.registrar(RegistrarTransicaoCommand.builder()
-                .sp(sp)
-                .tipo(TipoTransicao.MAPA_DISPONIBILIZADO)
-                .origem(admin)
-                .destino(sp.getUnidade())
-                .usuario(usuario)
-                .observacoes(request.observacoes())
-                .build());
-    }
-
-    private void validarMapaParaDisponibilizacao(Subprocesso subprocesso) {
-        Long codMapa = subprocesso.getMapa().getCodigo();
-        var competencias = mapaManutencaoService.buscarCompetenciasPorCodMapa(codMapa);
-
-        if (competencias.stream().anyMatch(c -> c.getAtividades().isEmpty())) {
-            throw new ErroValidacao("Todas as competências devem estar associadas a pelo menos uma atividade.");
-        }
-
-        var atividadesDoSubprocesso = mapaManutencaoService.buscarAtividadesPorMapaCodigo(codMapa);
-        var atividadesAssociadas = competencias.stream()
-                .flatMap(c -> c.getAtividades().stream())
-                .map(Atividade::getCodigo)
-                .collect(Collectors.toSet());
-
-        var atividadesNaoAssociadas = atividadesDoSubprocesso.stream()
-                .filter(a -> !atividadesAssociadas.contains(a.getCodigo()))
-                .toList();
-
-        if (!atividadesNaoAssociadas.isEmpty()) {
-            String nomesAtividades = atividadesNaoAssociadas.stream()
-                    .map(Atividade::getDescricao)
-                    .collect(Collectors.joining(", "));
-
-            throw new ErroValidacao(
-                    "Todas as atividades devem estar associadas a pelo menos uma competência.%nAtividades pendentes: %s"
-                            .formatted(nomesAtividades));
-        }
-    }
-
-    @Transactional
-    public void apresentarSugestoes(Long codSubprocesso, @Nullable String sugestoes, Usuario usuario) {
-        Subprocesso sp = workflowService.buscarSubprocesso(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp, MAPEAMENTO_MAPA_DISPONIBILIZADO, REVISAO_MAPA_DISPONIBILIZADO);
-
-        sp.getMapa().setSugestoes(sugestoes);
-        sp.setSituacao(SITUACAO_MAPA_COM_SUGESTOES.get(sp.getProcesso().getTipo()));
-
-        sp.setDataFimEtapa2(LocalDateTime.now());
-        subprocessoRepo.save(sp);
-
-        analiseFacade.removerPorSubprocesso(sp.getCodigo());
-
-        Unidade destino = sp.getUnidade().getUnidadeSuperior();
-        if (destino == null) {
-            destino = sp.getUnidade();
-        }
-
-        transicaoService.registrar(RegistrarTransicaoCommand.builder()
-                .sp(sp)
-                .tipo(TipoTransicao.MAPA_SUGESTOES_APRESENTADAS)
-                .origem(sp.getUnidade())
-                .destino(destino)
-                .usuario(usuario)
-                .observacoes(sugestoes)
-                .build());
-    }
-
-    @Transactional
-    public void validarMapa(Long codSubprocesso, Usuario usuario) {
-        Subprocesso sp = workflowService.buscarSubprocesso(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp, MAPEAMENTO_MAPA_DISPONIBILIZADO, REVISAO_MAPA_DISPONIBILIZADO);
-
-        sp.setSituacao(SITUACAO_MAPA_VALIDADO.get(sp.getProcesso().getTipo()));
-
-        sp.setDataFimEtapa2(LocalDateTime.now());
-        subprocessoRepo.save(sp);
-
-        Unidade destino = sp.getUnidade().getUnidadeSuperior();
-        if (destino == null) {
-            destino = sp.getUnidade();
-        }
-
-        transicaoService.registrar(RegistrarTransicaoCommand.builder()
-                .sp(sp)
-                .tipo(TipoTransicao.MAPA_VALIDADO)
-                .origem(sp.getUnidade())
-                .destino(destino)
-                .usuario(usuario)
-                .build());
-    }
-
-    @Transactional
-    public void devolverValidacao(Long codSubprocesso, @Nullable String justificativa, Usuario usuario) {
-        Subprocesso sp = workflowService.buscarSubprocesso(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp,
-                MAPEAMENTO_MAPA_COM_SUGESTOES, MAPEAMENTO_MAPA_VALIDADO,
-                REVISAO_MAPA_COM_SUGESTOES, REVISAO_MAPA_VALIDADO);
-
-        SituacaoSubprocesso novaSituacao = SITUACAO_MAPA_DISPONIBILIZADO.get(sp.getProcesso().getTipo());
-        sp.setDataFimEtapa2(null);
-
-        Unidade unidadeSuperior = sp.getUnidade().getUnidadeSuperior();
-        if (unidadeSuperior == null) {
-            unidadeSuperior = sp.getUnidade();
-        }
-
-        transicaoService.registrarAnaliseETransicao(RegistrarWorkflowCommand.builder()
-                .sp(sp)
-                .novaSituacao(novaSituacao)
-                .tipoTransicao(TipoTransicao.MAPA_VALIDACAO_DEVOLVIDA)
-                .tipoAnalise(TipoAnalise.VALIDACAO)
-                .tipoAcaoAnalise(TipoAcaoAnalise.DEVOLUCAO_MAPEAMENTO)
-                .unidadeAnalise(unidadeSuperior)
-                .unidadeOrigemTransicao(unidadeSuperior)
-                .unidadeDestinoTransicao(sp.getUnidade())
-                .usuario(usuario)
-                .motivoAnalise(justificativa)
-                .observacoes(justificativa)
-                .build());
-    }
-
-    @Transactional
-    public void aceitarValidacao(Long codSubprocesso, Usuario usuario) {
-        executarAceiteValidacao(codSubprocesso, usuario);
-    }
-
-    private void executarAceiteValidacao(Long codSubprocesso, Usuario usuario) {
-        Subprocesso sp = workflowService.buscarSubprocesso(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp,
-                MAPEAMENTO_MAPA_COM_SUGESTOES, MAPEAMENTO_MAPA_VALIDADO,
-                REVISAO_MAPA_COM_SUGESTOES, REVISAO_MAPA_VALIDADO);
-
-        Unidade unidadeAtual = obterUnidadeLocalizacao(sp);
-        Unidade proximaUnidade = unidadeAtual.getUnidadeSuperior();
-
-        // Se não tem próxima unidade (é o topo ou estrutura rasa), homologar direto
-        if (proximaUnidade == null) {
-            // Caso especial: Fim da cadeia de validação (Homologação Implícita?)
-            Unidade sup = unidadeAtual.getUnidadeSuperior();
-            String siglaUnidade = sup != null ? sup.getSigla() : unidadeAtual.getSigla();
-
-            analiseFacade.criarAnalise(sp, CriarAnaliseCommand.builder()
-                    .codSubprocesso(codSubprocesso)
-                    .observacoes("Aceite da validação")
-                    .tipo(TipoAnalise.VALIDACAO)
-                    .acao(TipoAcaoAnalise.ACEITE_MAPEAMENTO)
-                    .siglaUnidade(siglaUnidade)
-                    .tituloUsuario(usuario.getTituloEleitoral())
-                    .motivo(null)
-                    .build());
-
-            sp.setSituacao(SITUACAO_MAPA_HOMOLOGADO.get(sp.getProcesso().getTipo()));
-            subprocessoRepo.save(sp);
-        } else {
-            SituacaoSubprocesso novaSituacao = SITUACAO_MAPA_VALIDADO.get(sp.getProcesso().getTipo());
-            transicaoService.registrarAnaliseETransicao(RegistrarWorkflowCommand.builder()
-                    .sp(sp)
-                    .novaSituacao(novaSituacao)
-                    .tipoTransicao(TipoTransicao.MAPA_VALIDACAO_ACEITA)
-                    .tipoAnalise(TipoAnalise.VALIDACAO)
-                    .tipoAcaoAnalise(TipoAcaoAnalise.ACEITE_MAPEAMENTO)
-                    .unidadeAnalise(unidadeAtual)
-                    .unidadeOrigemTransicao(unidadeAtual)
-                    .unidadeDestinoTransicao(proximaUnidade)
-                    .usuario(usuario)
-                    .motivoAnalise("Aceite da validação")
-                    .build());
-        }
-    }
-
-    private Unidade obterUnidadeLocalizacao(Subprocesso sp) {
-        if (sp.getCodigo() == null) {
-            return sp.getUnidade();
-        }
-        return movimentacaoRepo.findFirstBySubprocessoCodigoOrderByDataHoraDesc(sp.getCodigo()).filter(m -> m.getUnidadeDestino() != null).map(Movimentacao::getUnidadeDestino).orElse(sp.getUnidade());
-    }
-
-    @Transactional
-    public void homologarValidacao(Long codSubprocesso, Usuario usuario) {
-        executarHomologacaoValidacao(codSubprocesso, usuario);
-    }
-
-    private void executarHomologacaoValidacao(Long codSubprocesso, Usuario usuario) {
-        Subprocesso sp = workflowService.buscarSubprocesso(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp,
-                MAPEAMENTO_MAPA_COM_SUGESTOES, MAPEAMENTO_MAPA_VALIDADO,
-                REVISAO_MAPA_COM_SUGESTOES, REVISAO_MAPA_VALIDADO);
-
-        sp.setSituacao(SITUACAO_MAPA_HOMOLOGADO.get(sp.getProcesso().getTipo()));
-        subprocessoRepo.save(sp);
-
-        Unidade admin = organizacaoFacade.buscarEntidadePorSigla(SIGLA_ADMIN);
-        transicaoService.registrar(RegistrarTransicaoCommand.builder()
-                .sp(sp)
-                .tipo(TipoTransicao.MAPA_HOMOLOGADO)
-                .origem(admin)
-                .destino(admin)
-                .usuario(usuario)
-                .build());
-    }
-
-    @Transactional
-    public void submeterMapaAjustado(Long codSubprocesso, SubmeterMapaAjustadoRequest request, Usuario usuario) {
-        Subprocesso sp = workflowService.buscarSubprocesso(codSubprocesso);
-        workflowService.validarSituacaoPermitida(sp, REVISAO_CADASTRO_HOMOLOGADA, REVISAO_MAPA_AJUSTADO);
-        workflowService.validarAssociacoesMapa(sp.getMapa().getCodigo());
-
-        sp.setSituacao(SITUACAO_MAPA_DISPONIBILIZADO.get(sp.getProcesso().getTipo()));
-        sp.setDataFimEtapa1(LocalDateTime.now());
-
-        if (request.dataLimiteEtapa2() != null) {
-            sp.setDataLimiteEtapa2(request.dataLimiteEtapa2());
-        }
-
-        subprocessoRepo.save(sp);
-
-        transicaoService.registrar(RegistrarTransicaoCommand.builder()
-                .sp(sp)
-                .tipo(TipoTransicao.MAPA_DISPONIBILIZADO)
-                .origem(sp.getUnidade())
-                .destino(sp.getUnidade())
-                .usuario(usuario)
-                .observacoes(request.justificativa())
-                .build());
-    }
-
-    @Transactional
-    public void disponibilizarMapaEmBloco(List<Long> subprocessoCodigos,
-            DisponibilizarMapaRequest request, Usuario usuario) {
-        subprocessoCodigos.forEach(codSubprocesso -> executarDisponibilizacaoMapa(codSubprocesso, request, usuario));
-    }
-
-    @Transactional
-    public void aceitarValidacaoEmBloco(List<Long> subprocessoCodigos, Usuario usuario) {
-        subprocessoCodigos.forEach(codSubprocesso -> executarAceiteValidacao(codSubprocesso, usuario));
-    }
-
-    @Transactional
-    public void homologarValidacaoEmBloco(List<Long> subprocessoCodigos, Usuario usuario) {
-        subprocessoCodigos.forEach(codSubprocesso -> executarHomologacaoValidacao(codSubprocesso, usuario));
-    }
-}
