+++ backend/src/test/java/sgc/subprocesso/service/SubprocessoServiceEmailTest.java
@@ -1,5 +1,6 @@
-package sgc.subprocesso.service.notificacao;
+package sgc.subprocesso.service;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -8,7 +9,14 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
+import sgc.alerta.AlertaFacade;
 import sgc.alerta.EmailService;
+import sgc.comum.model.ComumRepo;
+import sgc.subprocesso.dto.MapaAjusteMapper;
+import sgc.mapa.service.CopiaMapaService;
+import sgc.mapa.service.ImpactoMapaService;
+import sgc.mapa.service.MapaManutencaoService;
+import sgc.mapa.service.MapaSalvamentoService;
 import sgc.organizacao.OrganizacaoFacade;
 import sgc.organizacao.UsuarioFacade;
 import sgc.organizacao.dto.UnidadeResponsavelDto;
@@ -16,9 +24,9 @@ import sgc.organizacao.model.Unidade;
 import sgc.organizacao.model.Usuario;
 import sgc.processo.model.Processo;
 import sgc.processo.model.TipoProcesso;
-import sgc.subprocesso.model.Subprocesso;
-import sgc.subprocesso.model.TipoTransicao;
-import sgc.subprocesso.service.SubprocessoEmailService;
+import sgc.seguranca.SgcPermissionEvaluator;
+import sgc.subprocesso.dto.RegistrarTransicaoCommand;
+import sgc.subprocesso.model.*;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
@@ -28,20 +36,34 @@ import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-@DisplayName("SubprocessoEmailService Test")
-class SubprocessoEmailServiceTest {
+@DisplayName("SubprocessoService - Email Notifications")
+class SubprocessoServiceEmailTest {
 
-    @Mock
-    private EmailService emailService;
-    @Mock
-    private TemplateEngine templateEngine;
-    @Mock
-    private OrganizacaoFacade OrganizacaoFacade;
-    @Mock
-    private UsuarioFacade usuarioFacade;
+    @Mock private SubprocessoRepo subprocessoRepo;
+    @Mock private MovimentacaoRepo movimentacaoRepo;
+    @Mock private ComumRepo repo;
+    @Mock private AnaliseRepo analiseRepo;
+    @Mock private AlertaFacade alertaService;
+    @Mock private OrganizacaoFacade organizacaoFacade;
+    @Mock private UsuarioFacade usuarioFacade;
+    @Mock private ImpactoMapaService impactoMapaService;
+    @Mock private CopiaMapaService copiaMapaService;
+    @Mock private EmailService emailService;
+    @Mock private TemplateEngine templateEngine;
+    @Mock private MapaManutencaoService mapaManutencaoService;
+    @Mock private MapaSalvamentoService mapaSalvamentoService;
+    @Mock private MapaAjusteMapper mapaAjusteMapper;
+    @Mock private SgcPermissionEvaluator permissionEvaluator;
 
     @InjectMocks
-    private SubprocessoEmailService service;
+    private SubprocessoService service;
+
+    @BeforeEach
+    void setup() {
+        service.setMapaManutencaoService(mapaManutencaoService);
+        service.setSubprocessoRepo(subprocessoRepo);
+        service.setMovimentacaoRepo(movimentacaoRepo);
+    }
 
     @Test
     @DisplayName("Envia email para unidade destino")
@@ -54,7 +76,16 @@ class SubprocessoEmailServiceTest {
 
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DEVOLVIDO, new Unidade(), dest, "Motivo");
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DEVOLVIDO)
+                .origem(new Unidade())
+                .destino(dest)
+                .observacoes("Motivo")
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         verify(emailService).enviarEmailHtml(eq("dest@tre-pe.jus.br"), anyString(), eq("html"));
     }
@@ -78,9 +109,17 @@ class SubprocessoEmailServiceTest {
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
         // Mock responsavel para evitar NPE
-        when(OrganizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(mock(UnidadeResponsavelDto.class));
+        when(organizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(mock(UnidadeResponsavelDto.class));
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, origem, destino, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(origem)
+                .destino(destino)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         // Verifica envio para superior
         verify(emailService).enviarEmailHtml(eq("sup@tre-pe.jus.br"), anyString(), eq("html"));
@@ -97,27 +136,37 @@ class SubprocessoEmailServiceTest {
         // Ensure template engine is called, so exception is thrown in the try block
         lenient().when(templateEngine.process(anyString(), any(Context.class))).thenThrow(new RuntimeException("Template error"));
 
-        assertThatCode(() -> service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, null, dest, null))
-                .doesNotThrowAnyException();
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        assertThatCode(() -> service.registrarTransicao(cmd)).doesNotThrowAnyException();
     }
 
     @Test
     @DisplayName("Não envia email se tipo de transição não exige")
     void naoEnviaSeTipoNaoExige() {
-        service.notificarMovimentacao(null, TipoTransicao.CADASTRO_HOMOLOGADO, null, null, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(new Subprocesso())
+                .tipo(TipoTransicao.CADASTRO_HOMOLOGADO) // Does not send email? Checking TipoTransicao...
+                // Wait, CADASTRO_HOMOLOGADO might not send email.
+                // Assuming original test `naoEnviaSeTipoNaoExige` used this type.
+                .origem(new Unidade())
+                .destino(new Unidade())
+                .usuario(new Usuario())
+                .build();
+
+        // But cmd.sp() cannot be null as SubprocessoService uses it.
+        // And Movimentacao requires user.
+
+        service.registrarTransicao(cmd);
 
         verifyNoInteractions(emailService);
         verifyNoInteractions(templateEngine);
-    }
-
-    @Test
-    @DisplayName("Não envia email se destinatário nulo")
-    void naoEnviaSeDestinoNulo() {
-        Subprocesso sp = criarSubprocesso();
-        
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, null, null, null);
-
-        verifyNoInteractions(emailService);
     }
 
     @Test
@@ -129,7 +178,15 @@ class SubprocessoEmailServiceTest {
 
         lenient().when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, null, destino, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(null) // Null origin
+                .destino(destino)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         // Logs error because variable creation fails with null origin, so no email sent
         verifyNoInteractions(emailService);
@@ -153,14 +210,22 @@ class SubprocessoEmailServiceTest {
         destino.setSigla("DEST");
 
         // Mock responsavel para evitar NPE
-        lenient().when(OrganizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(mock(UnidadeResponsavelDto.class));
+        lenient().when(organizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(mock(UnidadeResponsavelDto.class));
 
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
         doNothing().when(emailService).enviarEmailHtml(eq("dest@tre-pe.jus.br"), anyString(), any());
         doThrow(new RuntimeException("Fail")).when(emailService).enviarEmailHtml(eq("sup@tre-pe.jus.br"), anyString(), any());
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, origem, destino, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(origem)
+                .destino(destino)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         verify(emailService).enviarEmailHtml(eq("sup@tre-pe.jus.br"), anyString(), any());
         verify(emailService).enviarEmailHtml(eq("dest@tre-pe.jus.br"), anyString(), any());
@@ -175,7 +240,16 @@ class SubprocessoEmailServiceTest {
 
         Unidade orig = new Unidade(); orig.setSigla("O");
         Unidade dest = new Unidade(); dest.setSigla("D");
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, orig, dest, null);
+
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(orig)
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
         verify(emailService).enviarEmailHtml(eq("d@tre-pe.jus.br"), anyString(), any());
     }
 
@@ -188,7 +262,16 @@ class SubprocessoEmailServiceTest {
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
         Unidade dest = new Unidade(); dest.setSigla("D");
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, new Unidade(), dest, null);
+
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         verify(templateEngine).process(eq("cadastro-disponibilizado"), argThat(ctx -> 
             ctx instanceof Context && ((Context)ctx).getVariable("dataLimiteEtapa2") != null
@@ -203,7 +286,17 @@ class SubprocessoEmailServiceTest {
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
         Unidade dest = new Unidade(); dest.setSigla("D");
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, new Unidade(), dest, "Minha Observação");
+
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .observacoes("Minha Observação")
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         verify(templateEngine).process(eq("cadastro-disponibilizado"), argThat(ctx ->
                 ctx instanceof Context && "Minha Observação".equals(((Context)ctx).getVariable("observacoes"))
@@ -218,7 +311,16 @@ class SubprocessoEmailServiceTest {
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
 
         Unidade dest = new Unidade(); dest.setSigla("D");
-        service.notificarMovimentacao(sp, TipoTransicao.PROCESSO_INICIADO, new Unidade(), dest, null);
+
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.PROCESSO_INICIADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         // Verifica que enviou email com assunto formatado
         verify(emailService).enviarEmailHtml(any(),
@@ -240,13 +342,21 @@ class SubprocessoEmailServiceTest {
                 .unidadeCodigo(1L)
                 .substitutoTitulo("123456")
                 .build();
-        when(OrganizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(resp);
+        when(organizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(resp);
 
         Usuario substituto = new Usuario();
         substituto.setEmail("sub@teste.com");
         when(usuarioFacade.buscarUsuarioPorTitulo("123456")).thenReturn(Optional.of(substituto));
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, new Unidade(), dest, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         verify(emailService).enviarEmailHtml(eq("sub@teste.com"), anyString(), eq("html"));
     }
@@ -265,13 +375,21 @@ class SubprocessoEmailServiceTest {
                 .unidadeCodigo(1L)
                 .substitutoTitulo("123456")
                 .build();
-        when(OrganizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(resp);
+        when(organizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(resp);
 
         Usuario substituto = new Usuario();
         substituto.setEmail(""); // Email vazio
         when(usuarioFacade.buscarUsuarioPorTitulo("123456")).thenReturn(Optional.of(substituto));
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, new Unidade(), dest, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         // Verifica envio para a unidade (padrão), mas NÃO para o substituto
         verify(emailService, times(1)).enviarEmailHtml(anyString(), anyString(), anyString());
@@ -287,9 +405,17 @@ class SubprocessoEmailServiceTest {
         dest.setSigla("DEST");
 
         when(templateEngine.process(anyString(), any(Context.class))).thenReturn("html");
-        when(OrganizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(null);
+        when(organizacaoFacade.buscarResponsavelUnidade(1L)).thenReturn(null);
 
-        service.notificarMovimentacao(sp, TipoTransicao.CADASTRO_DISPONIBILIZADO, new Unidade(), dest, null);
+        RegistrarTransicaoCommand cmd = RegistrarTransicaoCommand.builder()
+                .sp(sp)
+                .tipo(TipoTransicao.CADASTRO_DISPONIBILIZADO)
+                .origem(new Unidade())
+                .destino(dest)
+                .usuario(new Usuario())
+                .build();
+
+        service.registrarTransicao(cmd);
 
         // Apenas o email da unidade deve ser enviado
         verify(emailService, times(1)).enviarEmailHtml(anyString(), anyString(), anyString());
